pyproject.toml
[tool.poetry]
name = "sagira-bot"
version = "0.1.0"
description = "Discord Bot providing statistics and other tools for Destiny 2"
authors = ["Christopher Goes <ghostofgoes@gmail.com>"]
license = "MIT"
keywords = ["bot", "bungie", "destiny", "destiny2", "discord"]
readme = "README.md"
homepage = "https://github.com/GhostofGoes/sagira-bot"
repository = "https://github.com/GhostofGoes/sagira-bot"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Web Environment",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: Implementation :: CPython",
    "Intended Audience :: Developers",
    "Topic :: Games/Entertainment",
    "Topic :: Games/Entertainment :: First Person Shooters",
    "Topic :: Communications",
    "Topic :: Communications :: Chat",
    "Topic :: Internet",
    "Topic :: Internet :: WWW/HTTP",
    "Typing :: Typed",
    "Natural Language :: English",
]

[tools.poetry.urls]
"Changelog" = "https://github.com/GhostofGoes/sagira-bot/blob/main/CHANGELOG.md"
"Issue tracker" = "https://github.com/GhostofGoes/sagira-bot/issues"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = { git = "https://github.com/Rapptz/discord.py.git", rev = "4ce60b4cacaa7ab7fd6186f48c73bba6d09ad6a8" }
loguru = "^0.6.0"
uvloop = { version = "^0.16.0", markers = "sys_platform != 'win32'" }
# aiobungie = { version = "^0.2.5", allow-prereleases = true }
aiobungie = { git = "https://github.com/nxtlo/aiobungie.git", branch = "master" }
elasticsearch = {extras = ["async"], version = "^8.1.2"}

[tool.poetry.dev-dependencies]
black = "^22.1.0"
codespell = "^2.1.0"
flake8 = "^4.0.1"
flake8-bugbear = "^22.1.11"
flake8-import-order = "^0.18.1"
flake8-tidy-imports = "^4.6.0"
flake8-print = "^4.0.0"
flake8-builtins = "^1.5.3"
flake8-blind-except = "^0.2.0"
flake8-comprehensions = "^3.8.0"
flake8-debugger = "^4.0.0"
pep8-naming = "^0.12.1"
flake8-string-format = "^0.3.0"
flake8-broken-line = "^0.4.0"
flake8-mutable = "^1.2.0"
flake8-annotations = "^2.7.0"
vulture = "^2.3"
pre-commit = "~=2.17.0"
python-dotenv = "~=0.19.2"
taskipy = "~=1.10.1"
flake8-bandit = "^3.0.0"
mypy = "^0.942"
#pytest = "^7.0.1"
#pytest-randomly = "^3.11.0"
#pytest-mock = "^3.7.0"
#pytest-cov = "^3.0.0"
#Sphinx = "^4.4.0"
#furo = "^2022.3.4"

[tool.taskipy.tasks]
start = "python -m sagira"
precommit = "pre-commit install"
# TODO: fix mypy
lint = "task run_precommit && task vulture && task spelling && task bandit && task flake8"
run_precommit = "pre-commit run --all-files"
flake8 = "flake8 sagira"
mypy = "mypy sagira"
bandit = "bandit --recursive sagira"
vulture = "vulture --min-confidence 100 sagira"
spelling = "codespell"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
